+++ src/web/002_helpers.js
@@ -24,6 +24,8 @@ local.queryLinks = function queryLinks(links, query) {
 //   - rel: can take multiple values, space-separated, which are ANDed logically
 //   - rel: will ignore the preceding scheme and trailing slash on URI values
 //   - rel: items preceded by an exclamation-point (!) will invert (logical NOT)
+var uriTokenStart = '\\{[\\+\\#\\.\\/\\;\\?\\&]?';
+var uriTokenEnd = '(\\,|\\})';
 local.queryLink = function queryLink(link, query) {
 	for (var attr in query) {
 		if (attr == 'rel') {
@@ -41,7 +43,7 @@ local.queryLink = function queryLink(link, query) {
 		else {
 			if (typeof link[attr] == 'undefined') {
 				// Attribute not explicitly set -- is it present in the href as a URI token?
-				if (RegExp('\\{[^\\}]*'+attr+'[^\\{]*\\}','i').test(link.href) === true)
+				if (RegExp(uriTokenStart+attr+uriTokenEnd,'i').test(link.href) === true)
 					continue;
 				// Is the test value not falsey?
 				if (!!query[attr])
@@ -340,6 +342,37 @@ local.viaToUri = function(via) {
 		uri = via.map(function(proxy) {
 			return encode((proxy.proto.name||'http').toLowerCase() + '://' + proxy.hostname);
 		}).join('/');
+	}
+	return uri;
+};
+
+// EXPORTED
+// breaks a proxy URI into the different parts
+// eg 'httpl://0.page/httpl%3A%2F%2Ffoo%2Fhttpl%253A%252F%252Fmy_worker.js%252F'
+// -> ['httpl://0.page/', 'httpl://foo/', 'httpl://my_worker.js/']
+local.parseProxyUri = function(uri) {
+	var parts = [];
+	var re = /^http(l|s)?/;
+	while (re.exec(uri)) {
+		var end = uri.indexOf('http', 8); // skip 8 to skip 'http(l|s)://'. If just http://, will skip 1 too many, but that shouldnt matter
+		if (end == -1)
+			break;
+		parts.push(uri.slice(0, end));
+		uri = decodeURIComponent(uri.slice(end));
+	}
+	parts.push(uri);
+	return parts;
+};
+
+// EXPORTED
+// builds a proxy URI out of an array of parts
+// eg ['httpl://0.page/', 'httpl://foo/', 'httpl://my_worker.js/']
+// -> 'httpl://0.page/httpl%3A%2F%2Ffoo%2Fhttpl%253A%252F%252Fmy_worker.js%252F'
+local.makeProxyUri = function(parts) {
+	var uri = 0;
+	for (var i=parts.length-1; i >= 0; i--) {
+		if (!uri) uri = parts[i];
+		else uri = local.joinUri(parts[i], encodeURIComponent(uri));
 	}
 	return uri;
 };